Added full logging support
Added Named Pipes to enable logging into Touch Portal logs tab.
One named pipes where the core send the logs to the logger which will print them to the console, so in the logs tab.
The other one is to ensure that both apps (Logger and Core) are running, else they close.